backend/requirements.txt
PyJWT
python-multipart
passlib
bcrypt
python-dotenv
psycopg2-binary
pydantic
email-validator
